Object-cache & db: Change loading order to load from mu-plugins
